+++ packages/browser/types/index.d.ts
@@ -1,6 +1,29 @@
-import { History, HickoryLocation, PartialLocation, AnyLocation, Options as RootOptions } from '@hickory/root';
-export { History, HickoryLocation, PartialLocation, AnyLocation };
+import {
+  History,
+  Key,
+  Pathname,
+  Query,
+  Hash,
+  State,
+  LocationDetails,
+  PartialLocation,
+  HickoryLocation,
+  AnyLocation,
+  Options as RootOptions
+} from "@hickory/root";
+export {
+  History,
+  HickoryLocation,
+  PartialLocation,
+  AnyLocation,
+  LocationDetails,
+  Key,
+  Pathname,
+  Query,
+  Hash,
+  State
+};
 export interface Options extends RootOptions {
-    raw?: (pathname: string) => string;
+  raw?: (pathname: Pathname) => Pathname;
 }
 export default function Browser(options?: Options): History;