+++ packages/hash/types/index.d.ts
@@ -1,7 +1,30 @@
-import { History, HickoryLocation, PartialLocation, AnyLocation, Options as RootOptions } from '@hickory/root';
-export { History, HickoryLocation, PartialLocation, AnyLocation };
+import {
+  History,
+  Key,
+  Pathname,
+  Query,
+  Hash,
+  State,
+  LocationDetails,
+  HickoryLocation,
+  PartialLocation,
+  AnyLocation,
+  Options as RootOptions
+} from "@hickory/root";
+export {
+  History,
+  HickoryLocation,
+  PartialLocation,
+  AnyLocation,
+  Key,
+  Pathname,
+  Query,
+  Hash,
+  State,
+  LocationDetails
+};
 export interface Options extends RootOptions {
-    raw?: (pathname: string) => string;
-    hashType?: string;
+  raw?: (pathname: string) => string;
+  hashType?: string;
 }
-export default function Hash(options?: Options): History;
+export default function HashHistory(options?: Options): History;